feat: agent compile pipeline — PR review fixes
Address all review feedback from PR #102 across four tiers:

Critical:
- Restore compile_ws auth guards (membership + EditContent + own_branch)

High:
- Authenticate SSE agent event stream with membership guard + workspace filter
- MCP server auth via shared token (COWIKI_MCP_AUTH_TOKEN auto-generated)
- Workspace slug validation in gateway (block ../, /, \)
- Fix cowiki_search: correct table (pages) + workspace_slug filter
- Dual timeout: hard (idle→kill, 300s) + soft (total→warn, 1800s)
- Per-workspace concurrency cap via Semaphore map (default 4)

Medium:
- Wire stop_agent: store pid in process map for kill support
- Delete dead code: crates/db/src/embed.rs, CompileDrawer.tsx
- Log DB upsert failures instead of silently discarding indexing errors
- Validate embedder batch response length + zero-vector guard
- Fix SSE reconnect: fetch+ReadableStream replaces EventSource
- Update docs/mcp.md for rmcp architecture

Low:
- Standardize COWIKI_MCP_PORT as port-only (bind 127.0.0.1 unconditionally)
- TokenUsage u32→u64 to prevent overflow
- Remove dead CompileMode, keep CompileRequest minimal (branch only)
- Downgrade source listing log from warn! to debug!
- Add API key auth to /api/usage endpoint

+++ crates/core/src/ai/llm/openai.rs
@@ -143,11 +143,11 @@ impl Llm for OpenAILlm {
 
         let usage = resp.usage.map(|u| {
             if let Ok(mut tracker) = self.tracker.lock() {
-                tracker.record(&self.config.model, u.prompt_tokens, u.completion_tokens);
+                tracker.record(&self.config.model, u.prompt_tokens as u64, u.completion_tokens as u64);
             }
 
             let mut tu = TokenUsage::default();
-            tu.update(u.prompt_tokens, u.completion_tokens);
+            tu.update(u.prompt_tokens as u64, u.completion_tokens as u64);
             tu
         });
